api/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Task Queue
celery[redis]==5.3.6
redis==5.0.1

# HTTP
httpx==0.26.0
aiohttp==3.9.3

# Crypto Analysis
cryptography==42.0.2
pyOpenSSL==24.0.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
networkx==3.2.1

# AI/ML
torch==2.2.0
torch-geometric==2.5.0
scikit-learn==1.4.0

# Reporting
jinja2==3.1.3
reportlab==4.1.0
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
orjson==3.9.13
tenacity==8.2.3
apscheduler==3.10.4

# WebSocket
websockets==12.0

# CORS
starlette==0.36.3
